# First pass at optimizing SplitOnWordBoundaries