## Motivation and Context
@Velfi observed this function consuming a significant amount of codegen time.

## Description
Optimize split on word boundaries, primarily by avoiding recomputation of `isStartOfWord` when we already know it isn't the start of a word.

## Testing
- Original: 3.4s. Updated: 688ms
- Existing tests are exhaustive

On a `:aws:sdk:assemble` for the smoke test services, this improved from 53 seconds to 42 seconds
----

_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice._